+++ src/info.py
@@ -1,4 +1,4 @@
 class Info:
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.25b1" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2024-10-21"
+    AppVersionStr = "0.25b2" # see https://peps.python.org/pep-0440/
+    AppDateStr = "2025-02-22"